+++ commonroad/tests/visualization/test_icons.py
@@ -16,7 +16,7 @@ from commonroad.visualization.mp_renderer import MPRenderer
 
 __author__ = "Simon Sagmeister"
 __copyright__ = "TUM Cyber-Physical Systems Group"
-__version__ = "2021.4"
+__version__ = "2022.1"
 __email__ = "simon.sagmeister@tum.de"
 __status__ = "Released"